Dark current measurements added
Dark current measurements can now be taken by pressing a button on the front panel. Amplifier settings are written out to an indicator on screen (but not saved to disk). 

Dark current measurements proceed by resetting the amplifier input offset settings to default and bringing the dark current to within 5e-5 V of zero with a tolerance/error of 5e-5 V. 

The shutter is also included in the program and closes for dark current measurements and opens for data acquisition.